Make the default int an i64 (#3428)
* Make the default int an i64

* fmt

* Fix random integer

* Treat pids as i64 for now

+++ crates/nu-command/src/commands/length.rs
@@ -94,7 +94,7 @@ impl Iterator for CountIterator {
             input.count()
         };
 
-        Some(UntaggedValue::int(length).into_value(self.tag.clone()))
+        Some(UntaggedValue::int(length as i64).into_value(self.tag.clone()))
     }
 }
